requirements.txt
reflex==0.8.28.post1
pandas
scikit-learn
requests
python-dotenv
firebase-admin
razorpay